Simplify simulate_single_particle()
the function was first implemented to be as general as possible,
the loop checked at each iteration which barrier the particle would
collide with. Since non linear barriers were never implemented we can be
sure that the particle will never bouce with the same barrier two times
in a row, hence the changes.